reorganizes into component and hook folders
